+++ internal/balance/entity/withdrawal.go
@@ -6,11 +6,11 @@ import (
 
 type Withdrawal struct {
 	Order       string    `json:"order"`
-	Sum         int       `json:"sum"`
+	Sum         float32   `json:"sum"`
 	ProcessedAt time.Time `json:"processed_at"`
 }
 
 type WithdrawRequest struct {
-	Order string `json:"order" binding:"required"`
-	Sum   int    `json:"sum" binding:"required"`
+	Order string  `json:"order" binding:"required"`
+	Sum   float32 `json:"sum" binding:"required"`
 }
