+++ internal/orders/entity/order.go
@@ -9,7 +9,7 @@ import (
 type Order struct {
 	Number     string    `json:"number"`
 	Status     string    `json:"status"`
-	Accrual    int       `json:"accrual"`
+	Accrual    float32   `json:"accrual"`
 	UploadedAt time.Time `json:"uploaded_at"`
 	UserID     uuid.UUID `json:"-"`
 }
